requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==39.0.0
Django==4.1.5
django-autoslug==1.9.8
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-ckeditor==6.5.1
django-extensions==3.2.1
django-js-asset==2.0.0
django-timezone-field==5.0
idna==3.4
kombu==5.2.4
msgpack==1.0.4
Pillow==9.4.0
prompt-toolkit==3.0.36
pycparser==2.21
pydot==1.4.2
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7.1
razorpay==1.3.0
redis==4.4.2
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
trycourier==4.4.0
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
